From Marc Beyer:
Hi,

loading the wifty HomePage in Konqueror (v. 3.3.2 on Debian Sarge)
brings up a JavaScript error:

http://p6:8888/static/js/rico.js: TypeError: Value undefined (result of
expression document.attachEvent) is not an object. Cannot be called.

My JavaScript knowledge is poor to nonexistant, so I have no idea
whether this is an error in rico.js or Konqueror, but the attached patch
(stolen shamelessly from jamest@ajrs.com at
http://trac.openrico.org:9004/rico/ticket/2) prevents the warning.

Hope that helps,

Marc

git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@773 e84bef0a-9b06-0410-84ba-c4c9edb13aeb