Change the accessor generation for overriden accessors to use _foo_accessor() rather than declare _foo and wrap it with foo().  This is the preferred way to override Class::Accessor accessors and avoids having to special case each overriden accessor.
Take the list of properties in the Jifty::Web::Form::Field POD and make them
real accessors.  I'm not 100% sure they should all be declared, but some
are documented yet undeclared.  constructor is declared and documented in
Jifty::Param so maybe it should be removed from Field.  Anyhow, better to
have a maintained list of accessors than an unmaintained comment.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@2123 e84bef0a-9b06-0410-84ba-c4c9edb13aeb