From:     evdb@ecclestoad.co.uk     Subject:    Re: [jifty-devel] 'smart_tag_wrapper' to make creating smarter tags easier     Date:   March 19, 2007 11:36:20 AM EDT     To:       jifty-devel@lists.jifty.org     Reply-To:     jifty-devel@lists.jifty.org
On 19/03/07, Jesse Vincent <jesse@bestpractical.com> wrote:
I needed this tonight. Wanna make it go? ;)
On Mar 16, 2007, at 5:26 AM, Edmund von der Burg wrote:

I'd love to make it go but can't at the moment as I am not able to
connect to the net this week (except via a local libary machine that
is so crippled that it does not count).

The code at the start of this thread for 'form' works for me, please
add it as you see fit :)

Cheers,
 Edmund

> Hello,
>
> I've just committed code to Template::Declare that should make it
> easier to create smarter wrappers in the templates - for example the
> 'form' wrapper. Please see the diff for the details.
>
>  http://svn.jifty.org/index.cgi/jifty/revision/?rev=2989
>
> Using this the 'form' sub in J::V::Declare::Helpers would become:
>
> {
>    no warnings qw/redefine/;
>    sub form (&) {
>        my $code = shift;
>
>        smart_tag_wrapper {
>          outs_raw( Jifty->web->form->start(@_) );
>          $code->();
>          outs_raw( Jifty->web->form->end );
>          return '';
>        };
>    }
> }
>
>
> you could then call it like this (not tested properly):
>
> my $action = ...setup_action..;
> with( call => $next, name => "loginbox" ),

git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@3028 e84bef0a-9b06-0410-84ba-c4c9edb13aeb