A move to hand on the rendering of JDBI::Collection columns to app develoeprs.
We used to just ignore the rendering of any columns that refers to
a JDBI::Collection, because there's no trivial way to do so. 

However, if the column comes with a "render_as", we can assume that
the app developer know what he/she is doing.  So we just render it
as whatever specified.

One scenario is that a collection of tags is rendered as a normal
text input field, and develoeprs create tags records in their Tag
model in before_create() method. With ajax canonicalize_tags this
could be a simple and smart interface, and not difficult to implement
at all.

I'm sure that it'll open the chance to let developers write their
own collection rendering modules like I did before. If we keep
being ignoring, this won't even happend.


git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@3305 e84bef0a-9b06-0410-84ba-c4c9edb13aeb