Close Net::Server client sockets on background
HTTP::Server::Simple, when Net::Server implementations are used, does
not store away the created Net::Server subclasses.  This means that it
is not possible to find and close the client connections after a
fork().  We amend the Net::Server::Fork and Net::Server::PreFork
classes to store themselves on the Jifty::Server object, so we can
close client connections as need be later.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@6927 e84bef0a-9b06-0410-84ba-c4c9edb13aeb