In forking servers, disconnect the database before stalling on accept
In forking servers, the main server process never accesses the
database, so it does not need to hold a database connection open.
After working, ither it or its children need to set InactiveDestroy,
or it needs to not have a connection at all.  Since a connection is
not needed, we choose the latter.

git-svn-id: svn+ssh://svn.bestpractical.com/svn/jifty.org/jifty/trunk@6955 e84bef0a-9b06-0410-84ba-c4c9edb13aeb