Fold t/encoding-2.t into t/encoding.t
The t/encoding-2.t test was only failing because Jifty::Test::Email is
based on Test::Email, which is based on MIME::Entity -- which makes no
attempt to parse or understand character sets.  Thus, the message
failed to match the regular expression, which was in characters.

Using the same infrastructure as t/encoding.t causes the test to
succeed.