Update Jifty::JSON to use modern JSON.pm (2.xx)
Jifty::JSON's pure Perl functions no longer worked with modern
JSON.pm's, and JSON::Syck has been more or less deprecated in favor of
modern JSON and JSON::XS.

This deprecates the "singlequote" option that the previous version of
Jifty::JSON supported.  Single quoted strings are a violation of the
JSON spec.  Instead, HTML escape the entire blob of JSON if you plan to
put it in an HTML attribute.  Jifty::JSON will croak if you attempt to
use the "singlequote" option.