Revert a bunch of HTML escaping in <script> tags
It broke pages in the browser.  Although we serve content with an XHTML
doctype, the Content-Type is always text/html.  This forces the browser
to parse <script> tags according to the HTML spec (no entity parsing in
inline JS) instead of the XHTML spec (parsed entities).

This is not an ideal solution, but the least bad one for now.