Check continual path against canonicalized version, as URI::Escape and URI disagree on "safe" characters
One would expect that the following statement should be code should be
silent for all values of $char:

    use URI;
    use URI::Escape;

    my $uri = URI->new("http://localhost/".uri_escape($char));
    print "oh no!" unless $uri->as_string eq $uri->canonical->as_string;

That is, that URI::Escape and URI would agree on the set of characters
that need to be escaped.  Unfortunately, URI::Escape defaults to
escaping as per RFC3986, whereas URI uses the less restrictive
RFC2732.  The latter allows the following characters which the former
does not:

    ! * ' ( )

Plack::Request returns a ->canonical'd version of the request URI,
which means that one cannot assume that the URI that a user was
redirected to is the same as the URI that they later arrive at.  A
"more correct" solution would be to either not canonicalize the URI in
Plack::Request, or to change all URI::Escape::uri_escape calls to
escape only RFC2732-unfriendly characters; however, this fix is
arguably also correct, and much simpler to implement.