Keep the current plack env when restoring from a continuation
This avoids having to specifically swap in various psgi and plack
variables, as we were doing manually.  It also means that the
request's ->referer, ->remote_host, etc are correct for the _current_
request, rather than being empty (if the request was hand-constructed)
or being from the stored request.

It is slightly overzealous in that ->parameters will disagree with
->env->{QUERY_STRING}, for instance, but ->parameters is correctly
inherited from the request in the continuation.