Only cache the current_user object if the session is loaded
This prevents negative lookups (due to the session not having been
loaded yet) from being cached indefinitely.  Plugins which do things
early (in ->new_request, or the 'have_request' callback) will still
get an empty current_user, but this will no longer be incorrectly
cached for the duration of the request.