Make the CAS do its own serving, as opposed to relying on CCJS
This paves the way for a more generic CAS server, allowing non-CCJS to
be served.

It also removes the 'cdn' option, which was of limited use because it
required a complete app mirror.  The more correct solution involves
either:
 a) publishing other static content (images, etc) to the CDN, and
    rewriting the CSS to point to them
 b) rewriting the CSS to point to the application's standard images

Either of these solutions, since they involve parsing and rewriting CSS,
is also relevant to hosting a Jifty application not at the server root.