Make using Jifty::Test::WWW::Declare warnings free
The warnings pragma doesn't cross package boundaries, and Exporter was
still throwing a redefined warning for get().  Sub::Exporter squashes
this warning, so use that instead.  It adds a little more complexity to
our already multi-headed import method, but not much.  Despite the added
dep in Makefile.PL, Sub::Exporter should already exist thanks to
pre-existing deps.

We may just want to remove the declarative mech test framework entirely
since it's not really used for Jifty, but it's possible some apps use
it.