Only load view handlers and actions once
This speeds up startup, particularly under multi-threaded plack servers,
by allowing the view handlers and actions to be loaded once before
forking.  This is safe because essentially no state is saved (and almost
certainly none is modified) on the Jifty::Handler and Jifty::API objects
between first and second initialization.