Allow a shortcut around the dispatcher for fragments
Re-dispatching through the application's dispatcher can be a significant
performance hit on pageregion-heavy pages.  Allow dispatchers to declare
a fragment_handler method which will be used in place of the full
dispatcher.

Care must be taken to ensure that this does not allow walking around
ACLs.  Anything which runs on every request (sessions, Jifty->api
limiting) will have already run once on the original
/__jifty/webservices/json request; however, since that page is in no way
ACL protected by the dispatcher, a fragment_handler method which does
not adequately express the ACL checks of the rest of the dispatcher is a
security vulnerability.  Whitelisting, rather than blacklisting, is most
likely the correct course of action.