Canonicalize all request paths; this catches fragment requests as well
Previously, the path as passed in the fragment request data structure
was used verbatim in the dispatcher and other locations.  This possibly
allowed requests to walk around ACLs by requesting
'/some/safe/place/../../../dangerous' as a fragment.  As a non-fragment,
this would have been canonicalized to '/dangerous', but fragment paths
were not being so canonicalized.