Checking for path hygene in continuation returns is no longer necessary
This warning was introduced in 36bd338, when "call" and "return" were
handled by the same code.  The only way we determined which phase we
were in was by whether the path matched what was stored in the
continuation.  The danger is this was that the path that we redirected
to could be different from the path that the browser then requested, if
it chose to escape characters differently; this would cause Jifty to
redirect indefinitely.  The warning was placed here to warn of the
perils of redirecting to URIs containing characters which the browser
might choose to escape differently than Jifty did.

However, the canonicalization done by Plack::Request (and in 3e54cf2)
guarantees that the URIs have been canonicalized equivilently; thus,
this warning is always a false-positive.