Ensure that Template::Declare halos run in the <body>
When 8d11269 removed the mostly-unnecessary "proscribed" abstraction, it
removed the logic which ensured that Template::Declare halos only ran
inside of the <body> tag; HTML::Mason halos were unaffected, as they use
a separate codepath.

Restore the logic that ensures that Template::Declare halos are only
rendered where it is appropriate.