Short-circuit halo close on error
end_component_hook is still called for every component even if the
request is in the midst of aborting (because of a run-time error, for
instance).  Detect such cases and short-circuit the closing halo code,
which would simply be discarded.  Note that $context->error is always a
scalar ref (either to an empty string, or an exception object), as Mason
allows the plugin to modify the error.