Split apply_fragment_updates into XML parsing and application
apply_fragment_updates consists of two halves -- walking an XML DOM to
extract response information, and applying that response information to
update the fragment on the page.  Split the function into
xml_fragment_updates and fragment_updates, so that non-XML codepaths can
reuse the latter function to apply changes to fragments.