Add a basic Web::Hippie::Pipe endpoint at /__jifty/_hippie
Using the Jifty->bus AnyMQ object as its base, construct a
Web::Hippie::Pipe endpoint at /__jifty/_hippie.  This establishes
bi-directional real-time communication with the client, using wither
websockets, mxhr (multi-part XMLHttpRequest), or simply polling --
whichever the browser supports.

This can be used to send messages to the Hippie.Pipe object in the
browser, by subscribing $listener to topics on new connections (when
$path is "/new_listener").  Messages recieved from the browser are
presented as requests to "/message", and disconnects (when the browser
moves to a new page) are presented as requests to "/error".

For now, simply present a no-op endpoint.