Add a Jifty AnyMQ subclass which pushes "type" explicitly
If a client subscribes to multiple topics, they have no way of
differentiating which topic a newly-arrived message came from; as a
convention, Jifty uses the "type" key of the data to store the topic
that the message was published on.  Provide an AnyMQ subclass to catch
data which already uses that key, and to provide an interface which
papers over the call to ->topic().