Intentionally ignore subscriptions during region push
When regions were rendered in response to events, they had the
opportunity to attempt to add new subscriptions via Jifty->subs->add.
However, such calls added to the in-memory store for the client_id that
had already been fetched -- and thus were lost.  Despite this, two
different pieces of code attempted to manage the region subscriptions
during event-triggered region rendering.

Remove the misleading code, and replace it with an explicit mention of
the missing functionality.